refactored code as a python module, replaced pathlib with os.path
